+++ OpenTabletDriver.Configurations/Configurations/VEIKK/VK640.json
@@ -23,7 +23,7 @@
       "ProductID": 5,
       "InputReportLength": 13,
       "OutputReportLength": 9,
-      "ReportParser": "OpenTabletDriver.Configurations.Parsers.Veikk.VeikkTiltedReportParser",
+      "ReportParser": "OpenTabletDriver.Configurations.Parsers.Veikk.VeikkTiltReportParser",
       "FeatureInitReport": null,
       "OutputInitReport": [
         "CQEE",
